refactor(aco): clarify code-defined FLP naming
Rename `FlpProvider.getFlps()` to `getCodeFlps()` so it is clear the
method returns code-defined FLPs collected from registered `FlpFactory`
instances, and not folder-level permissions in general.

Rename `UpdateFolderUseCaseWithoutInheritedPermissions` to
`UpdateFolderUseCaseWithoutReadOnlyPermissions` — the decorator strips
both inherited and code-defined permissions, so the old name only
described half of what it does.
